pyproject.toml
[tool.poetry]
name = "mathservice"
version = "0.1.0"
description = ""
authors = ["Jong012 <Jong012>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = ">=3.2,<4.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
django-dotenv = "^1.4.2"
pytest = "^7.2.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.8.1"
django-extensions = "^3.2.1"
ipython = "^8.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
